requirements.txt
asgiref==3.5.2
Django==4.0.4
gunicorn==20.1.0
psycopg2==2.8.6
sqlparse==0.4.2
whitenoise==6.2.0
Pillow==9.4.0
django-storages==1.13.2
boto3==1.26.154
python-decouple==3.8
django-location-field==2.7.1